Update jp.m3u
m3u links for the following domains are not available



[vnpt.nekocdn.xyz] ← The URL is distributed under a different domain, but if someone posts the URL on another playlist on github, the link will likely spread and cause a surge in server access, so we have been asked not to publish it, so we will not post it here. Please do not post it here.

[r3jx.djtmewibu.com] 

The link for this domain was an experimental test distribution and has already been discontinued. It is not expected to be restored.    

cdns.jp-primehome.com] 

The link for this domain is the source of a server of a distributor that offers Japanese TV channels to Japanese residents outside Japan for a fee.

It was supposed to be playable only through a dedicated viewing app provided by the distributor for subscribers, but someone noticed that the URLs of the per-channel links in the app were not tokenized, and posted the per-channel links on github.

At first, only a few people who saw the github link accessed the app, but off-the-record information about accessing the paid service link for free was spread by YouTubers in Japan through their videos, and access to jp-primehome skyrocketed! This caused a sudden increase in access to jp-primehome, and it became known to the distributors that github was the cause of the unauthorized access. In October 2024, all link URLs were tokenized by the distributor, and now, a random number code is added to the existing URLs to unlock the access restriction, otherwise the connection cannot be established.

Even if this random number code is obtained from somewhere, the same code cannot be shared among multiple people at the same time.

If another person tries to connect using the same code, a connection error will occur and the connection will be denied.